requirements.txt
django-markdownify
django-allauth==0.45.0
importlib-metadata==4.13.0
django
django-crispy-forms
martor
sentry-sdk
gunicorn
python-dotenv
dj-database-url
psycopg2; platform_system == "Windows"
psycopg2-binary; sys_platform == "linux"
